// src/app/data/savefile-expanded/sections/World/currentScriptProgress.d.ts
export interface CurrentScriptProgress {
    [key: string]: number

    oaksLab: number
    paletteTown: number
    rivalsHouse: number
    viridianCity: number
    pewterCity: number
    route3: number
    route4: number
    viridianGym: number
    pewterGym: number
    ceruleanGym: number
    vermillionGym: number
    celadonGym: number
    route6: number
    route8: number
    route24: number
    route25: number
    route9: number
    route10: number
    mtMoon1: number
    mtMoon3: number
    ssAnne8: number
    ssAnne9: number
    route22: number
    playersHouse2: number
    viridianMarket: number
    route22Gate: number
    ceruleanCity: number
    ssAnne5: number
    viridianForest: number
    museum1: number
    route13: number
    route14: number
    route17: number
    route19: number
    route21: number
    safariZoneEntrance: number
    rockTunnel2: number
    rockTunnel1: number
    route11: number
    route12: number
    route15: number
    route16: number
    route18: number
    route20: number
    ssAnne10: number
    vermillionCity: number
    pokemonTower2: number
    pokemonTower3: number
    pokemonTower4: number
    pokemonTower5: number
    pokemonTower6: number
    pokemonTower7: number
    rocketHideout1: number
    rocketHideout2: number
    rocketHideout3: number
    rocketHideout4: number
    route6Gate: number
    route8Gate: number
    cinnabarIsland: number
    mansion1: number
    mansion2: number
    mansion3: number
    mansion4: number
    victoryRoad2: number
    victoryRoad3: number
    fightingDojo: number
    silphCo2: number
    silphCo3: number
    silphCo4: number
    silphCo5: number
    silphCo6: number
    silphCo7: number
    silphCo8: number
    silphCo9: number
    hofRoom: number
    rival: number
    lorelei: number
    bruno: number
    agatha: number
    unknownDungeon3: number
    victoryRoad1: number
    lance: number
    silphCo10: number
    silphCo11: number
    fuchsiaGym: number
    saffronGym: number
    cinnabarGym: number
    celadonGameCorner: number
    route16Gate: number
    billsHouse: number
    route5Gate: number
    powerPlantRoute7Gate: number
    ssAnne2: number
    seafoamIslands4: number
    route23: number
    seafoamIslands5: number
    route18Gate: number
}